Moved GetClamps into PatchEPhys, renamed some variables in IVCurve, tested working
